Add RPG Roles markdown file for collaboration
Hello @willemlarsen, thank you for the wonderful game.

After facilitating this at the Mob Programming Conference this year with Chris, I thought it would be good to get a plain-text file that we can do pull requests on to discuss ideas. I'm open to other ways of collaborating if you have ideas!